+++ .gitignore
@@ -1,2 +1,3 @@
 templates/
 .vscode/
+__pycache__/
