+++ csv_parse.log
@@ -1,0 +1,8 @@
+2021-11-15 16:07:57,265 - INFO - 130716 - ------------------------------------------------------------------------------------------------------------
+2021-11-15 16:07:57,266 - INFO - 130716 - Outlier detected in Feb.22 for date 2021-09-21 06:58:00: value -16.55 and zscore 100
+2021-11-15 16:07:57,515 - INFO - 130716 - ------------------------------------------------------------------------------------------------------------
+2021-11-15 16:07:57,515 - INFO - 130716 - Outlier detected in Nov.21 for date 2021-09-24 21:10:00: value -7.95 and zscore 100
+2021-11-15 16:07:57,515 - INFO - 130716 - ------------------------------------------------------------------------------------------------------------
+2021-11-15 16:07:57,515 - INFO - 130716 - Outlier detected in Dec.21 for date 2021-09-24 21:10:00: value -9.75 and zscore 100
+2021-11-15 16:07:57,522 - INFO - 130716 - ------------------------------------------------------------------------------------------------------------
+2021-11-15 16:07:57,523 - INFO - 130716 - Outlier detected in Dec.21 for date 2021-09-22 06:13:00: value -0.25 and zscore 100
